style: Rename template-flow-modal to flow-template-modal and update imports
