+++ web/components/flow/canvas-modal/index.ts
@@ -2,4 +2,4 @@ export * from './add-flow-variable-modal';
 export * from './export-flow-modal';
 export * from './import-flow-modal';
 export * from './save-flow-modal';
-export * from './template-flow-modal';
+export * from './flow-template-modal';
